Added legend to web dev + more
Kyle added a legend to the web development grid so that it is easier to keep track of all the colors that are being used to represent things.

Also, Kyle made an interactive map for Jose that contains the 170x25 dimensions and shows all the locations of the sensors within the grid that are not "broken"

Kyle also fixed the coloring scheme for the localization guesses. Previously, there was not a brown spot for each grouping of localization guesses, but now there is and it is the correct one. The brown spot is the one the localization algorithm thinks is most likely the emitter location.

Also fixed a minor bug in the JSON file where there was an extra column of data that was just numbers. We don't need that so I removed it.