WebDev, df500 to Jose
Did some work converting files like block_500_only into Jose format because he wants more data. This results in a lot of csv files being produced, so might have to make a folder for all of it within the datasets folder

Did some work with the web development stuff. Got the 170x25 grid to show on screen but lost the localization points, which is fine at this point.