Multiple Grids Visualization
This will be Kyle's last push to do with coding updates. The next couple updates will have to do with documentation and preparing to share the code with others.

Created a function in GridLib.py that can create multiple grids on a single map when you give it the parameters of a grid structure.

Fixed a minor bug in DataFiltrationLib.py where the sensor_trim() function returned a dataframe with an extra index column. Now there is only one index column.

Added new constants to Constants.py. We now have a constant for the origins of grids that cover street blocks 500, 400, 300, and 200 on Clematis Street. Left older version of block 500 constant in there because we have been using it and I don't want to go through all our code to replace it.

Created a new file called mobIntelProcess.py that shows how to use gl.multipleGrids and also created a giant grid to see how many tiles we would create. Also created a loop that shows how many probes a probe request burst has. This was determined by rows in mobintel data that had the same machash and probingtime. It is very slow and only should be used for understanding purposes.

Added in a bunch of pictures for archives